+++ src/bin/esyPrintEnv.js
@@ -4,7 +4,7 @@
 
 import type {CommandContext} from './esy';
 
-import {getBuildSandbox, getBuildConfig} from './esy';
+import {getSandbox, getBuildConfig} from './esy';
 import * as Task from '../build-task';
 import * as Sandbox from '../sandbox';
 import * as Env from '../environment';
@@ -13,9 +13,9 @@ export default async function esyPrintEnv(ctx: CommandContext) {
   // TODO: It's just a status command. Print the command that would be
   // used to setup the environment along with status of
   // the build processes, staleness, package validity etc.
-  const sandbox = await getBuildSandbox(ctx);
+  const sandbox = await getSandbox(ctx);
   const config = await getBuildConfig(ctx);
-  const task = Task.fromBuildSandbox(sandbox, config);
+  const task = Task.fromSandbox(sandbox, config);
   const env = Sandbox.getCommandEnv(task, config);
   console.log(Env.printEnvironment(task.env));
 }
