+++ src/builders/makefile-builder.js
@@ -2,13 +2,7 @@
  * @flow
  */
 
-import type {
-  BuildSpec,
-  BuildTask,
-  BuildTaskCommand,
-  Config,
-  BuildSandbox,
-} from '../types';
+import type {BuildSpec, BuildTask, BuildTaskCommand, Config, Sandbox} from '../types';
 
 import {sync as mkdirp} from 'mkdirp';
 import createLogger from 'debug';
@@ -16,7 +10,7 @@ import outdent from 'outdent';
 
 import * as Graph from '../graph';
 import * as Task from '../build-task';
-import * as Sandbox from '../sandbox';
+import * as S from '../sandbox';
 import * as Env from '../environment';
 import * as Makefile from '../Makefile';
 import {normalizePackageName} from '../util';
@@ -194,11 +188,7 @@ const defineSandboxEnvRule = Makefile.createDefine({
 /**
  * Render `build` as Makefile (+ related files) into the supplied `outputPath`.
  */
-export function eject(
-  sandbox: BuildSandbox,
-  outputPath: string,
-  config: Config<path.Path>,
-) {
+export function eject(sandbox: Sandbox, outputPath: string, config: Config<path.Path>) {
   const buildFiles = [];
 
   function generateMetaRule({filename}) {
@@ -335,7 +325,7 @@ export function eject(
 
   // Emit build artefacts for packages
   log('process dependency graph');
-  const rootTask = Task.fromBuildSandbox(sandbox, config);
+  const rootTask = Task.fromSandbox(sandbox, config);
 
   const finalInstallPathSetMetaRule = generateMetaRule({
     filename: 'final-install-path-set.txt',
@@ -413,7 +403,7 @@ export function eject(
         export ESY_EJECT__STORE=$(esyGetStorePathFromPrefix "$HOME/.esy")
       fi
 
-      ${Env.printEnvironment(Sandbox.getSandboxEnv(rootTask, config))}
+      ${Env.printEnvironment(S.getSandboxEnv(rootTask, config))}
     `,
   };
